tinkerpopperise step names for more consistency: hasLabel/hasId (#108)
label(x) -> hasLabel(x)
id(x) -> hasId(x)

I changed my opinion on these because the latter versions are more
consistent with e.g. the property filter steps: `property(key)` is
a property lookup step, and not a filter step, which is why we have the
`hasXX` steps in the first place.