repeat step: allow both specific and generic steps, both in combination with behaviour builder (#114)
This introduces a new type parameter for repeat: `B >: A`, where `A` is the type where the Traversal starts (e.g. `Node`), and `B` is a generalisation of it (e.g. `Node`). It allows us to combine domain-specific and generic graph steps while keeping full type safety.

The `behaviourBuilder` moved into a second parameter list, so that the compiler can derive the type parameter `B`. 
Making that parameter implicit is not because it's supposed to be used implicitly, but rather allow the user to omit additional `()` when they want to stick with the default. 

In essence: it's using combination of small feats of the scala language, but it's all for the simplicity and correctness of the use-side :)