remove 'traversal already exhausted' warning (#125)
The idea was to warn the user that a traversal cannot be executed twice,
but unfortunately I didn't find a better place than to put this check in
`hasNext`.
There are plenty of legitimate reasons to check `hasNext` more than once,
leading to many of these warnings being logged e.g. during
codepropertygraph/test, which only causes confusion.