change where/not semantics (#132)
* was: `A => Traversal[_]`
* new: `Traversal[A] => Traversal[_]`

Should have been that way from the start - note e.g. that the tests
didn't need any update. prior to this change, it was assumed that the
user provides a query that is (implicitly) convertible into a Traversal.
We're making this explicit now, i.e. less magic on the usage side.