make node removal persistent (#139)
Prior to this, removing a node would apply to the current in-memory
graph, but it would reappear after closing and reopening the graph.
Two changes were needed to fix that:
1) explicitly remove it from the storage mvstore map
2) clear the `NodeRef.node` instance (of type `OdbNode`), so that the
reference manager cannot persist it. This is not the most obvious way
to achieve that, but the one with the smallest impact, since we don't
need to add another field like 'removed' to NodeRef, and can keep the
ReferenceManager's collection type as is.