Change API use for List-valued properties
This changes how we set properties: We used to unroll List-valued properties; now we send them at once. This change is necessary for performance, because we now store CONTAINED_NODE as properties -- and therefore we get long lists of properties. There are necessary downstream changes -- this patch is compatible with older codepropertygraph versions.