Make RepeatStep iterator reusable.
Factored out the iterator class used in RepeatStep and made it
independent of Traversal class.
Furthermore i moved the work list creation into the iterator class to
make sure that no funny stuff can happen because of the shared work
list. The way it is currently used works but if the generated traversals
are not completely drained before the next traversal is created and
used, the work items get mixed up.