Make @Doc a java annotation instead of scala annotation. (#237)
This enables us remove the scala-compiler dependency which currently
blocks building this project with scala 3.

Since the java annotations require us to specify the name of the
annotation parameters, this is a breaking API change. On the plus side
@help.Traversal and @Doc are now both java annotations and thus now
have a homogenious usage style.

I did not try to go for scala 3 macros to keep the change minimal and to
avoid having different code path for scala 2 and scala 3 for the time we
are cross building this project.