Add toSetMutable end step to Traversal.
Add toSetMutable end step to Traversal and use it in tests instead of
toSet.
This becomes necessary because the Scala 3 compiler cannot choose
between toSet defined on our Traversal which has signature
def toSet: mutable.Set[A]
and the toSet of IterableOnceExtensionMethods which has signature
def toSet[B >: A]: immutable.Set[B]
when used in combination with the "should" matchers of scalatest.

In order to not break API we leave the toSet methods as they are and
just change our test code.