downgrade to scala 3.3 LTS (#440)
* downgrade to scala 3.3 LTS

Projects that are used as libraries should stick to the latest LTS version.

Scala 3 guarantees backward compatibility across minor releases in the entire 3.x series, but not forward compatibility. This means that libraries compiled with any Scala 3.x version can be used in projects compiled with any Scala 3.y version with y >= x.

https://www.scala-lang.org/blog/2022/08/17/long-term-compatibility-plans.html#library-maintainers
https://docs.scala-lang.org/overviews/core/binary-compatibility-of-scala-releases.html

* fix actions